Initial Commit – Late
After hacking away at this for a few hours, I decided to check this into
version control. So far, I have a functional counter and a way to
associate the right coordinates with it. As I'm getting further along,
I keep butting up against the limits of not doing this with a stricter
OO process. So this is the current version. Next step is to refactor
the icon methods as objects that hold their own coordinates.