Incrementable and Decremental Counter
This commit builds a working counter object with a centered counter
value. It also successfully implements the prototypal inheritance set up
in the last commit by fixing a typo in the world prototype :poop:.

Lastly, it normalizes the objects' draw methods so that there's a nearly
consistent default canvas setting.